refactor(scaffold): move Label, FpsOverlay, RendererInfoProbe (#120 step 4)
Three pure-generic primitives moved verbatim out of
exhibits/quadrics/ into the scaffold layer:
  - Label.ts → scaffold/ui/Label.ts (two-line billboarded text)
  - FpsOverlay.ts → scaffold/perf/FpsOverlay.ts
  - RendererInfoProbe.ts → scaffold/perf/RendererInfoProbe.ts

No code changes inside the moved files; only the import path at the
single quadrics call site flips to @/scaffold/{ui,perf}/...
