+++ src/scaffold/perf/RendererInfoProbe.ts
